wsprRxTx

+++ python/wsprRxTx.py
@@ -73,11 +73,12 @@ pwr='20'
 freq=0
 tx=True
 VERSION='1.1'
-PROGRAM=sys.argv[0]
+PROGRAM=sys.argv[0].split(".")[0]
 cycle=5
-lckFile=("%s.lck") % PROGRAM.split(".")[0]
-logFile=("%s.log") % PROGRAM.split(".")[0]
-tlmFile=("%s.tlm") % PROGRAM.split(".")[0]
+lckFile=("%s.lck") % PROGRAM
+logFile=("%s.log") % PROGRAM
+tlmFile=("%s.tlm") % PROGRAM
+pidFile=("%s.pid") % PROGRAM
 myPID=os.getpid()
 DEBUGLEVEL=0
 #*------------------------------------------------------------------------
@@ -95,10 +96,51 @@ def createLock():
        f.write("%d\n" % myPID)
 
 def resetLock():
-    if os.path.exists(lckFile):
+    if isLock()==True:
        os.remove(lckFile)
     else:
        log(0,"The file %s does not exist" % lckFile)
+
+
+def isLock():
+    if os.path.exists(lckFile):
+       return True
+    else:
+       return False 
+
+def createPID():
+    log(0,"Archivo PID es %s" % pidFile)
+    with open(pidFile, 'a') as f:
+       f.write("%d\n" % myPID)
+
+def resetPID():
+    if isPID()!= True:
+       os.remove(pidFile)
+    else:
+       log(0,"The file %s does not exist" % pidFile)
+
+
+def isPID():
+    if os.path.exists(pidFile):
+       return True
+    else:
+       return False
+#*--------------------------------------------------------------------------
+#* isRunning()
+#*
+#*--------------------------------------------------------------------------
+def isRunning():
+    if isPID() == True:
+       f = open(pidFile, "r")
+       for line in f:
+           line=line.replace("\n","")
+           log(1,"Returned line from pidFile %s" % line)
+           return(line)
+    else:
+       log(0,"The % file does not exist" % pidFile)
+       return ""
+
+
 #*------------------------------------------------------------------------
 #* setPTT
 #* Activate the PTT thru GPIO27
@@ -124,26 +166,25 @@ def getFreq(b):
 #* print log thru standard output
 #*-------------------------------------------------------------------------
 def log(d,st):
-    m=datetime.datetime.now()
-    if logFile != '' : 
-       with open(logFile, 'a') as f:
-         #f.write(text)   
-         f.write('%s %s\n' % (m,st))
+    if d<=DEBUGLEVEL:
+       m=datetime.datetime.now()
+       if logFile != '' : 
+          with open(logFile, 'a') as f:
+               f.write('%s %s\n' % (m,st))
        print('%s %s' % (m,st))
 #*----------------------------------------------------------------------------
 #* Exception and Termination handler
 #*----------------------------------------------------------------------------
 def signal_handler(sig, frame):
    log(0,"signal_handler: WSPR Monitor and Beacon is being terminated, clean up completed!")
-   log(0,'Turning GPIO(27) low as PTT')
-   #GPIO.output(27, GPIO.LOW)
+   log(1,'Turning GPIO(27) low as PTT')
    setPTT(False)
 
    try:
-     log(0,"Process terminated, clean up completed!")
+     log(1,"Process terminated, clean up completed!")
      resetLock()
    except:
-     log(0,"signal_handler: Process finalized")
+     log(1,"signal_handler: Process finalized")
    sys.exit(0)
 #*-------------------------------------------------------------------------
 #* Exception management
@@ -151,86 +192,12 @@ def signal_handler(sig, frame):
 signal.signal(signal.SIGINT, signal_handler)
 signal.signal(signal.SIGTERM, signal_handler)
 
-#*----------------------------------------------------------------------------
-#* Exception and Termination handler
-#*----------------------------------------------------------------------------
-#def killProc(p):#
-#
-#   try:
-#     parent = psutil.Process(p.pid)
-#     log(2,"killProc: process %s)" % parent)
-#     for child in parent.children(recursive=True):  # or parent.children() for recursive=False
-#       log(2,"killProc:     -- child(%s)" % child)
-#       child.kill()
-#     parent.kill()
-#   except UnboundLocalError as error:
-#     log(0,"killProc: Unable to kill parent PID")
-#   except NameError as error:
-#     log(2,"killProc: Allocation error")
-#   except Exception as exception:
-#     pass
-#*-----------------------------------------------------------------------------------------------
-#* waitProc
-#* waiting for a process to start looking for a given string
-#*-----------------------------------------------------------------------------------------------
-#def waitProc(p,stOK,timeout):
-#    log(2,"waitProc: waiting for string(%s) for timeout(%d)" % (stOK,timeout))
-#    ts=time.time()
-#    while (time.time()-ts) <= timeout:
-#        s=non_block_read(p.stdout)
-#        if s!="":
-#           if s.find(stOK) != -1:
-#              log(2,"waitProc: process (%s) found OK string (%s)" % (str(psutil.Process(p.pid)),stOK))
-#              return 0
-#           else:
-#              log(1,s.replace("\n",""))
-#    return -1
-#*------------------------------------------------------------------------------------------------
-#def startProc(cmd,stOK):
-#    log(2,"startProc: cmd(%s) OK(%s)" % (cmd,stOK))
-#    p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-#    if waitProc(p,stOK,5) == 0:
-#       log(1,"startProc: Process successfully launched")
-#       return p
-#    log(0,"startProc: failed to launch process *ABORT*")
-#    raise Exception('startProc: general exceptions not caught by specific handling')
-#    return None
-#*-------------------------------------------------------------------------
-#* isProcess(string)
-#* Return PID of a process if exits 0 otherwise
-#*
-#*-------------------------------------------------------------------------
-#def isProcess(procName):
-#    cmd = 'sudo ps -aux | pgrep %s' % procName
-#    p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-#    for line in p.stdout.readlines():
-#        st=line.decode("utf-8")
-#        if st=="":
-#           return 0
-#        i=int(st)
-#        if i==os.getpid():
-#           pass
-#        else:
-#           return i
-#    return 0
-#    retval = p.wait()
-#*--------------------------------------------------------------------------
-#* isFile(filename)
-#* Check if a file exits
-#*--------------------------------------------------------------------------
-#def isFile(filename):
-#  try:
-#    os.stat(filename)
-#    return True
-#  except:
-#    return False
-#
 #*--------------------------------------------------------------------------
 #* doExec()
 #* Execute a command and return the result
 #*--------------------------------------------------------------------------
 def doExec(cmd):
-    log(0,"doExec: [cmd] %s" % cmd)
+    log(1,"doExec: [cmd] %s" % cmd)
     p = subprocess.Popen(cmd, shell=True, 
                          stdout=subprocess.PIPE,
                          stderr=subprocess.PIPE)
@@ -238,8 +205,6 @@ def doExec(cmd):
     st=''
     for line in p.stdout:
         st=st+line.decode("utf-8").replace("\n","")
-    #log(0,"doExec: [result] %s" % st)
-    #log(0,"doExec: [error] %s" % p.stderr)
     return st
 
 #*--------------------------------------------------------------------------
@@ -248,7 +213,7 @@ def doExec(cmd):
 #*--------------------------------------------------------------------------
 def doShell(cmd):
 
-    log(0,"doShell: [cmd] %s" % cmd)
+    log(1,"doShell: [cmd] %s" % cmd)
     p = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     (result, error) = p.communicate()
 
@@ -257,7 +222,6 @@ def doShell(cmd):
     if rc != 0:
         log(0,"Error: failed to execute command: %s" % cmd)
         log(0,error)
-    #log(0,"doShell: [Result] %s" %  result)
     return result
 
 #*--------------------------------------------------------------------------
@@ -266,16 +230,16 @@ def doShell(cmd):
 #*--------------------------------------------------------------------------
 def doService(freq):
 
-    log(0,'Executing time synchronization (ntpd sync)')
-    cmd='sudo /home/pi/ntpd.sync'
+    log(0,'Time Executing time synchronization')
+    cmd='sudo chronyc -a makestep'
+    #cmd='sudo /home/pi/ntpd.sync'
     result=doShell(cmd)
-    log(0,"doService: ntpd(%s)" % result)
+    log(0,"doService: TimeSync(%s)" % result)
  
     while True:
        #*--------------------------*
        #* PTT low (receive)        *
        #*--------------------------*
-       #GPIO.output(27, GPIO.LOW)
        setPTT(False)
 
        #*--------------------------*
@@ -285,7 +249,7 @@ def doService(freq):
        result=doShell(cmd)
        log(0,"[TL] %s" % result.replace("\n",""))
        with open(tlmFile, 'a') as tlm:
-          log(0,"writing telemetry to file %s" % tlmFile)
+          log(1,"writing telemetry to file %s" % tlmFile)
           tlm.write("%s" % result)
 
        #*--------------------------*
@@ -331,30 +295,15 @@ def startService():
     if freq==0 :
        log(0,'Non supported band(%s), exit' % band)
        exit()
-    log(0,"Starting daemon PID(%d) and creating %s lock" % (myPID,lckFile))
+    log(1,"Starting daemon PID(%d) and creating %s lock" % (myPID,lckFile))
     createLock()
-    log(0,"%s lock file created" % lckFile)
+    log(1,"%s lock file created" % lckFile)
     try:
       doService(freq)
     except:
       log(0,"Exception detected, program being terminated")
     else:
       log(0,"Program is ending normally")
-#*--------------------------------------------------------------------------
-#* isRunning()
-#*
-#*--------------------------------------------------------------------------
-def isRunning():
-    if os.path.exists(lckFile):
-       f = open(lckFile, "r")
-       for line in f:
-           line=line.replace("\n","")
-           log(0,"Returned line from lckFile %s" % line)
-           return(line)
-    else:
-       log(0,"The lockfile file does not exist")
-       return ""
-
 #*============================================================================
 #* Main Program
 #*
@@ -377,7 +326,7 @@ ap.add_argument("--band",help="Band to use i.e. 20m",required=False)
 ap.add_argument("--pwr",help="Set Power in dBm",required=False)
 ap.add_argument("--tx",help="Enable TX",required=False, action="store_true")
 ap.add_argument("--cycle",help="Cycles RX/TX",required=False)
-ap.add_argument("--log",help="Log activity to file",required=False,action="store_true")
+ap.add_argument("--log",help="Log activity to file",default=True,required=False,action="store_true")
 ap.add_argument("--lock",help="Lock further execution till reset",required=False,action="store_true")
 ap.add_argument("--reset",help="Reset locked execution", required=False,action="store_true")
 args=ap.parse_args()
@@ -390,7 +339,7 @@ if args.log == True :
    logFile="%s.log" % PROGRAM
    with open(logFile, 'w+') as f:
       f.write("")
-   log(0,'Set logfile(%s)' % logFile)
+   log(2,'Set logfile(%s)' % logFile)
 
 log(0,"Program %s Version %s PID(%d)" % (PROGRAM,VERSION,os.getpid()))
 
@@ -444,12 +393,15 @@ if args.pwr != None :
 #* Process start command     *
 #*---------------------------*
 if args.start:
+       if isLock() == True:
+          log(0,"Daemon is locked, run with --reset opetion to clear")
+          exit()
        PID=isRunning()
        if PID == '':
           setPTT(False)
           startService()
        else:
-          log(0,"Daemon already running PID(%s) or locked, exit" % PID)
+          log(0,"Daemon already running PID(%s), exit" % PID)
        exit()
 #*---------------------------*
 #* Process stop command      *
@@ -466,13 +418,18 @@ if args.stop:
        setPTT(False)
        exit()
 
-
+#*----------------------------------------------------------------------------
+#* Review lock status
+#*----------------------------------------------------------------------------
 if args.lock == True:
+       if isLock() == True:
+          log(0,"Daemon is already locked, exit")
+          exit()
        pid=isRunning()
        if(pid == ''):
          log(0,'Daemon is not running, locking and exit')
        else:
-         log(0,'Daemon found PID(%s), killing it' % pid)
+         log(0,'Daemon found PID(%s), killing, locking and exit' % pid)
          n=int(pid)
          cmd="sudo kill %s" % n
          result=doShell(cmd)
@@ -497,8 +454,8 @@ if args.reset == True:
 #*------------------------------------*
 pid=isRunning()
 if (pid == ''):
-    print(0,'Daemon is not running, exit')
+    print(0,'Status: Daemon is not running, exit')
 else:
-    print(0,'Daemon found PID(%s)' % pid)
+    print(0,'Status: Daemon is runnint PID(%s)' % pid)
 exit() 
 #*--------------------------------[End of program] -----------------------------------------------
